Iron-out circular option dependencies.
