+++ packages/ag-charts-community/src/chart/options/dropShadowOptions.ts
@@ -1,0 +1,14 @@
+import type { CssColor, PixelSize } from './types';
+
+export interface AgDropShadowOptions {
+    /** Whether or not the shadow is visible. */
+    enabled?: boolean;
+    /** The colour of the shadow. */
+    color?: CssColor;
+    /** The horizontal offset in pixels for the shadow. */
+    xOffset?: PixelSize;
+    /** The vertical offset in pixels for the shadow. */
+    yOffset?: PixelSize;
+    /** The radius of the shadow's blur, given in pixels. */
+    blur?: PixelSize;
+}
