+++ packages/ag-charts-community/src/chart/series/polar/polarTooltipOptions.ts
@@ -1,0 +1,17 @@
+import type { AgSeriesTooltipRendererParams } from '../../options/tooltipOptions';
+
+export interface AgPolarSeriesTooltipRendererParams extends AgSeriesTooltipRendererParams {
+    /** angleKey as specified on series options. */
+    readonly angleKey: string;
+    /** angleValue as read from series data via the angleKey property. */
+    readonly angleValue?: any;
+    /** angleName as specified on series options. */
+    readonly angleName?: string;
+
+    /** radiusKey as specified on series options. */
+    readonly radiusKey?: string;
+    /** radiusValue as read from series data via the radiusKey property. */
+    readonly radiusValue?: any;
+    /** radiusName as specified on series options. */
+    readonly radiusName?: string;
+}
